clean up global variables in some file change add and edit from dialog to actvity

+++ app/src/main/java/com/tresor/home/model/FinancialHistoryModel.java
@@ -36,7 +36,7 @@ public class FinancialHistoryModel implements Parcelable{
 
     protected FinancialHistoryModel(Parcel in) {
         amount = in.readString();
-        amountUnformatted = in.readInt();
+        amountUnformatted = in.readDouble();
         usesComma = in.readByte() != 0;
         currencyId = in.readInt();
         hashtag = in.createStringArrayList();
